+++ tty/pc-tty/ttypc.c
@@ -40,7 +40,7 @@ static void ttypc_poolthr(void *arg)
 	ttypc_t* ttypc = (ttypc_t *)arg;
 	const void *idata, *odata;
 	unsigned long req;
-	unsigned int rid;
+	unsigned long rid;
 	msg_t msg;
 	id_t id;
 	int err;
